+++ src/ByteDev.Collections/ArrayExtensions.cs
@@ -25,7 +25,7 @@ namespace ByteDev.Collections
             {
                 if (values == null)
                 {
-                    source[i] = default(TSource);
+                    source[i] = default;
                 }
                 else
                 {
@@ -36,6 +36,17 @@ namespace ByteDev.Collections
                     valuesIndex++;
                 }
             }
+        }
+
+        /// <summary>
+        /// Gets the total number of elements in the array. If the array is null then zero is returned.
+        /// </summary>
+        /// <typeparam name="TSource">Array element type.</typeparam>
+        /// <param name="source">The array to perform the operation on.</param>
+        /// <returns>Zero if the array is null; otherwise the array length.</returns>
+        public static int SafeLength<TSource>(this TSource[] source)
+        {
+            return source?.Length ?? 0;
         }
     }
 }